# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Golang BE API
  description: |
    Production-grade REST API dengan event-driven architecture.

    Stack: Gin, PostgreSQL (versioned migrations), Redis, Kafka (event streaming + DLQ),
    MongoDB (audit trail), Typesense (full-text search), Prometheus, OpenTelemetry.

    Envelope response mengikuti kontrak:
    `{ success, message, errorCode?, data, meta? }`.

    - `errorCode` — machine-readable code stabil (mis. `AUTH_TOKEN_EXPIRED`, `PRODUCT_NOT_FOUND`).
    - `meta` — pagination cursor untuk endpoint list.
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/v1/products:
    get:
      tags: [Products]
      summary: List my products (cursor-paginated)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: cursor
          in: query
          description: Opaque cursor from `meta.nextCursor` of the previous page. Omit for the first page.
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: OK — `meta.nextCursor` continues the listing while `meta.hasMore` is true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListEnvelope"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [Products]
      summary: Create product (publishes `product.created` to Kafka)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductEnvelope"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/products/catalog:
    get:
      tags: [Products]
      summary: Buyer marketplace catalog (all sellers, cursor-paginated)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListEnvelope"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/products/search:
    get:
      tags: [Products]
      summary: Full-text search (Typesense)
      description: Multi-match over name + description with fuzziness. 503 with errorCode `SERVICE_UNAVAILABLE` when search is disabled.
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Matching product documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchEnvelope"
        "503":
          $ref: "#/components/responses/Error"
  /api/v1/products/{id}:
    get:
      tags: [Products]
      summary: Product detail (Redis-cached)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductEnvelope"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [Products]
      summary: Update product (owner only, publishes `product.updated`)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductEnvelope"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Products]
      summary: Delete product (owner or admin, publishes `product.deleted`)
      description: Owners delete their own products. Users with role `admin` can delete any product (RBAC).
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /ws/products:
    get:
      tags: [Realtime]
      summary: WebSocket — real-time product event stream
      description: |
        Upgrade to WebSocket to receive `product.created` / `product.updated` / `product.deleted`
        events as they flow through Kafka (fan-out by the notifier consumer group).

        Auth via query param because browsers cannot set headers on WebSocket:
        `ws://localhost:8080/ws/products?token=<accessToken>`
      security: []
      responses:
        "101":
          description: Switching Protocols — event stream begins
        "401":
          description: Invalid token
components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: apikey
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error envelope with stable machine code
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data: {}
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: "not found" }
        errorCode:
          type: string
          description: Stable machine code
          enum:
            - RESOURCE_NOT_FOUND
            - VALIDATION_ERROR
            - AUTH_EMAIL_TAKEN
            - AUTH_TOKEN_EXPIRED
            - AUTH_UNAUTHORIZED
            - AUTH_INVALID_API_KEY
            - AUTH_FORBIDDEN
            - RATE_LIMIT_EXCEEDED
            - SERVICE_UNAVAILABLE
            - INTERNAL_ERROR
...
    Product:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        stock: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProductCreateRequest:
      type: object
      required: [name, price]
      properties:
        name: { type: string, example: "Kopi Susu" }
        description: { type: string, example: "Iced" }
        price: { type: number, example: 28000 }
        stock: { type: integer, example: 10 }
    ProductUpdateRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number }
        stock: { type: integer }
    ProductEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { $ref: "#/components/schemas/Product" }
    PageMeta:
      type: object
      properties:
        limit: { type: integer }
        nextCursor:
          type: integer
          description: Pass as `cursor` for the next page
        hasMore: { type: boolean }
    ProductListEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        meta: { $ref: "#/components/schemas/PageMeta" }
    SearchEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              userId: { type: integer }
              name: { type: string }
              description: { type: string }
              price: { type: number }